+++ backend/canisters/user_index/CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Remove all references to LocalGroupIndexes ([#8057](https://github.com/open-chat-labs/open-chat/pull/8057))
+- Return CHIT details in current user summary ([#8069](https://github.com/open-chat-labs/open-chat/pull/8069))
 
 ## [[2.0.1755](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1755-user_index)] - 2025-05-21